Make Linux version work
This commit makes the Linux version of the example work by proving all
`libQt*.so*` files (not only the ones ending in `.so`, but also
`.so.6`). Before, it didn't work as the libraries' `SONAME`s are
`*.so.6` and the Bazel rules only provided the `*.so` symlinks.

We also make `libicu*.so*` available, since all `libQt*.so*` libraries
depend on that.

Lastly, we put Qt's `plugin` directory into a `filegroup`, make that
available via the `data` argument, and set the
`QT_QPA_PLATFORM_PLUGIN_PATH` environment variable so that Qt finds it
at runtime.